Moved classes to under hivemall.mix.yarn
