+++ yarn/src/main/java/hivemall/mix/yarn/network/HeartbeatHandler.java
@@ -16,9 +16,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.mix.network;
+package hivemall.mix.yarn.network;
 
-import hivemall.utils.TimestampedValue;
+import hivemall.mix.yarn.utils.TimestampedValue;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
