+++ yarn/src/main/java/hivemall/mix/yarn/network/MixServerRequestHandler.java
@@ -16,10 +16,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.mix.network;
+package hivemall.mix.yarn.network;
 
 import hivemall.mix.yarn.MixYarnEnv;
-import hivemall.utils.TimestampedValue;
+import hivemall.mix.yarn.utils.TimestampedValue;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
